refactor: remove config manager
